refactor(service): LottoStore 을 책임에 맞게 서비스 나누기
하나의 클래스가 너무 많은 책임을 가지지 않기 위해 리팩토링
