+++ src/main/java/lotto/service/LottoResultService.java
@@ -1,0 +1,55 @@
+package lotto.service;
+
+import lotto.domain.Lotto;
+import lotto.domain.LottoRank;
+import lotto.domain.WinNumbers;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class LottoResultService {
+
+    public long sumAllPrize(List<Lotto> lottos, WinNumbers winNumbers) {
+        return getAllLottoResult(lottos,winNumbers).stream()
+                .mapToLong(r -> r.getPrize())
+                .sum();
+    }
+
+    private LottoRank getLottoResult(Lotto lotto, WinNumbers winNumbers) {
+        int correctCount = getMatchingCount(lotto, winNumbers);
+        boolean bonusMatching = isBonusMatching(lotto, winNumbers);
+        return LottoRank.getRank(correctCount, bonusMatching);
+    }
+
+    private List<LottoRank> getAllLottoResult(List<Lotto> lottos, WinNumbers winNumbers) {
+        List<LottoRank> lottoRanks = new ArrayList<>();
+
+        for (Lotto lotto : lottos) {
+            lottoRanks.add(getLottoResult(lotto, winNumbers));
+        }
+        return lottoRanks;
+    }
+
+    private boolean isBonusMatching(Lotto lotto, WinNumbers winNumbers) {
+        if (hasNum(lotto.getNumbers(), winNumbers.getBonusNum()))
+            return true;
+        return false;
+    }
+
+    private int getMatchingCount(Lotto lotto, WinNumbers winNumbers) {
+        int correctCount = 0;
+
+        for (int num : lotto.getNumbers()) {
+            if (hasNum(winNumbers.getWinNumber(), num)) {
+                correctCount++;
+            }
+        }
+        return correctCount;
+    }
+
+    private boolean hasNum(List<Integer> nums, int compareNum) {
+        if (nums.contains(compareNum))
+            return true;
+        return false;
+    }
+}
